style/#10: 네이버 코드 포매터 적용
