+++ src/main/java/com/campus/campus/domain/council/domain/entity/StudentCouncil.java
@@ -64,11 +64,14 @@ public class StudentCouncil extends BaseEntity {
 		this.password = newPassword;
 	}
 
-    public String getFullCouncilName() {
-        StringBuilder fullName = new StringBuilder();
-        if (school != null) fullName.append(school.getSchoolName());
-        if (college != null) fullName.append(" ").append(college.getCollegeName());
-        if (major != null) fullName.append(" ").append(major.getMajorName());
-        return fullName.append(" 학생회").toString().trim();
-    }
+	public String getFullCouncilName() {
+		StringBuilder fullName = new StringBuilder();
+		if (school != null)
+			fullName.append(school.getSchoolName());
+		if (college != null)
+			fullName.append(" ").append(college.getCollegeName());
+		if (major != null)
+			fullName.append(" ").append(major.getMajorName());
+		return fullName.append(" 학생회").toString().trim();
+	}
 }
