--- src/main/java/com/campus/campus/domain/studentcouncilpost/application/dto/NormalizedDateTime.java
@@ -1,8 +1,0 @@
-package com.campus.campus.domain.studentcouncilpost.application.dto;
-
-import java.time.LocalDateTime;
-
-public record NormalizedDateTime(
-        LocalDateTime startDateTime,
-        LocalDateTime endDateTime
-) {}
