+++ src/main/java/com/campus/campus/domain/studentcouncilpost/application/exception/NotPostWriterException.java
@@ -3,9 +3,9 @@ package com.campus.campus.domain.studentcouncilpost.application.exception;
 import com.campus.campus.global.common.exception.ApplicationException;
 
 public class NotPostWriterException extends ApplicationException {
-    public NotPostWriterException() {
-        super(ErrorCode.NOT_POST_WRITER);
-    }
+	public NotPostWriterException() {
+		super(ErrorCode.NOT_POST_WRITER);
+	}
 }
