+++ src/main/java/com/campus/campus/domain/studentcouncilpost/application/exception/PostImageLimitExceededException.java
@@ -3,7 +3,7 @@ package com.campus.campus.domain.studentcouncilpost.application.exception;
 import com.campus.campus.global.common.exception.ApplicationException;
 
 public class PostImageLimitExceededException extends ApplicationException {
-    public PostImageLimitExceededException() {
-        super(ErrorCode.POST_IMAGE_LIMIT_EXCEEDED);
-    }
+	public PostImageLimitExceededException() {
+		super(ErrorCode.POST_IMAGE_LIMIT_EXCEEDED);
+	}
 }
