+++ src/main/java/com/campus/campus/domain/studentcouncilpost/application/exception/PostNotFoundException.java
@@ -3,6 +3,8 @@ package com.campus.campus.domain.studentcouncilpost.application.exception;
 import com.campus.campus.global.common.exception.ApplicationException;
 
 public class PostNotFoundException extends ApplicationException {
-    public PostNotFoundException() {super(ErrorCode.POST_NOT_FOUND);}
+	public PostNotFoundException() {
+		super(ErrorCode.POST_NOT_FOUND);
+	}
 }
