+++ src/main/java/com/campus/campus/domain/studentcouncilpost/application/exception/ThumbnailRequiredException.java
@@ -3,8 +3,8 @@ package com.campus.campus.domain.studentcouncilpost.application.exception;
 import com.campus.campus.global.common.exception.ApplicationException;
 
 public class ThumbnailRequiredException extends ApplicationException {
-    public ThumbnailRequiredException() {
-        super(ErrorCode.THUMBNAIL_REQUIRED);
-    }
+	public ThumbnailRequiredException() {
+		super(ErrorCode.THUMBNAIL_REQUIRED);
+	}
 }
