+++ src/main/java/com/campus/campus/domain/studentcouncilpost/application/service/StudentCouncilPostService.java
@@ -1,12 +1,22 @@
 package com.campus.campus.domain.studentcouncilpost.application.service;
 
+import java.util.ArrayList;
+import java.util.List;
+
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
 import com.campus.campus.domain.council.application.exception.StudentCouncilNotFoundException;
 import com.campus.campus.domain.council.domain.entity.StudentCouncil;
 import com.campus.campus.domain.council.domain.repository.StudentCouncilRepository;
-import com.campus.campus.domain.studentcouncilpost.application.dto.NormalizedDateTime;
-import com.campus.campus.domain.studentcouncilpost.application.dto.PostListItemResponseDto;
-import com.campus.campus.domain.studentcouncilpost.application.dto.PostRequestDto;
-import com.campus.campus.domain.studentcouncilpost.application.dto.PostResponseDto;
+import com.campus.campus.domain.studentcouncilpost.application.dto.response.NormalizedDateTime;
+import com.campus.campus.domain.studentcouncilpost.application.dto.response.PostListItemResponseDto;
+import com.campus.campus.domain.studentcouncilpost.application.dto.request.PostRequestDto;
+import com.campus.campus.domain.studentcouncilpost.application.dto.response.PostResponseDto;
 import com.campus.campus.domain.studentcouncilpost.application.exception.NotPostWriterException;
 import com.campus.campus.domain.studentcouncilpost.application.exception.PostImageLimitExceededException;
 import com.campus.campus.domain.studentcouncilpost.application.exception.PostNotFoundException;
@@ -18,245 +28,238 @@ import com.campus.campus.domain.studentcouncilpost.domain.entity.StudentCouncilP
 import com.campus.campus.domain.studentcouncilpost.domain.repository.PostImageRepository;
 import com.campus.campus.domain.studentcouncilpost.domain.repository.StudentCouncilPostRepository;
 import com.campus.campus.global.oci.application.service.PresignedUrlService;
-import java.util.ArrayList;
-import java.util.List;
+
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
 public class StudentCouncilPostService {
 
-    private final StudentCouncilPostRepository postRepository;
-    private final StudentCouncilRepository studentCouncilRepository;
-    private final PostImageRepository postImageRepository;
-    private final PresignedUrlService presignedUrlService;
+	private final StudentCouncilPostRepository postRepository;
+	private final StudentCouncilRepository studentCouncilRepository;
+	private final PostImageRepository postImageRepository;
+	private final PresignedUrlService presignedUrlService;
 
-    @Transactional
-    public PostResponseDto create(Long councilId, PostRequestDto dto) {
+	@Transactional
+	public PostResponseDto create(Long councilId, PostRequestDto dto) {
 
-        if (dto.imageUrls() != null && dto.imageUrls().size() > 10) {
-            throw new PostImageLimitExceededException();
-        }
+		if (dto.imageUrls() != null && dto.imageUrls().size() > 10) {
+			throw new PostImageLimitExceededException();
+		}
 
-        StudentCouncil writer = studentCouncilRepository.findByIdWithDetails(councilId)
-                .orElseThrow(StudentCouncilNotFoundException::new);
+		StudentCouncil writer = studentCouncilRepository.findByIdWithDetails(councilId)
+			.orElseThrow(StudentCouncilNotFoundException::new);
 
-        if (dto.thumbnailImageUrl() == null && dto.thumbnailIcon() == null) {
-            throw new ThumbnailRequiredException();
-        }
+		if (dto.thumbnailImageUrl() == null && dto.thumbnailIcon() == null) {
+			throw new ThumbnailRequiredException();
+		}
 
-        NormalizedDateTime normalized = dto.category().validateAndNormalize(dto);
+		NormalizedDateTime normalized = dto.category().validateAndNormalize(dto);
 
-        StudentCouncilPost post = StudentCouncilPostMapper.toEntity(
-                writer,
-                dto,
-                normalized.startDateTime(),
-                normalized.endDateTime()
-        );
+		StudentCouncilPost post = StudentCouncilPostMapper.toEntity(
+			writer,
+			dto,
+			normalized.startDateTime(),
+			normalized.endDateTime()
+		);
 
-        postRepository.save(post);
+		postRepository.save(post);
 
-        if (dto.imageUrls() != null) {
-            for (String imageUrl : dto.imageUrls()) {
-                postImageRepository.save(
-                        StudentCouncilPostMapper.toEntity(post, imageUrl)
-                );
-            }
-        }
+		if (dto.imageUrls() != null) {
+			for (String imageUrl : dto.imageUrls()) {
+				postImageRepository.save(
+					StudentCouncilPostMapper.toEntity(post, imageUrl)
+				);
+			}
+		}
 
-        List<String> imageUrls = postImageRepository
-                .findAllByPostOrderByIdAsc(post)
-                .stream()
-                .map(PostImage::getImageUrl)
-                .toList();
+		List<String> imageUrls = postImageRepository
+			.findAllByPostOrderByIdAsc(post)
+			.stream()
+			.map(PostImage::getImageUrl)
+			.toList();
 
-        return StudentCouncilPostMapper.toDetail(
-                post,
-                imageUrls,
-                councilId
-        );
-    }
+		return StudentCouncilPostMapper.toDetail(
+			post,
+			imageUrls,
+			councilId
+		);
+	}
 
-    @Transactional(readOnly = true)
-    public PostResponseDto findById(Long postId, Long currentUserId) {
-        StudentCouncilPost post = postRepository.findByIdWithFullInfo(postId)
-                .orElseThrow(PostNotFoundException::new);
+	@Transactional(readOnly = true)
+	public PostResponseDto findById(Long postId, Long currentUserId) {
+		StudentCouncilPost post = postRepository.findByIdWithFullInfo(postId)
+			.orElseThrow(PostNotFoundException::new);
 
-        List<String> imageUrls = postImageRepository
-                .findAllByPostOrderByIdAsc(post)
-                .stream()
-                .map(PostImage::getImageUrl)
-                .toList();
+		List<String> imageUrls = postImageRepository
+			.findAllByPostOrderByIdAsc(post)
+			.stream()
+			.map(PostImage::getImageUrl)
+			.toList();
 
-        return StudentCouncilPostMapper.toDetail(
-                post,
-                imageUrls,
-                currentUserId
-        );
-    }
+		return StudentCouncilPostMapper.toDetail(
+			post,
+			imageUrls,
+			currentUserId
+		);
+	}
 
-    @Transactional(readOnly = true)
-    public Page<PostListItemResponseDto> findAll(
-            PostCategory category,
-            int page,
-            int size,
-            Long currentUserId
-    ) {
-        Pageable pageable = PageRequest.of(
-                Math.max(page - 1, 0),
-                size,
-                Sort.by(Sort.Direction.DESC, "createdAt")
-        );
+	@Transactional(readOnly = true)
+	public Page<PostListItemResponseDto> findAll(
+		PostCategory category,
+		int page,
+		int size,
+		Long currentUserId
+	) {
+		Pageable pageable = PageRequest.of(
+			Math.max(page - 1, 0),
+			size,
+			Sort.by(Sort.Direction.DESC, "createdAt")
+		);
 
-        Page<StudentCouncilPost> posts = (category == null)
-                ? postRepository.findAll(pageable)
-                : postRepository.findAllByCategory(category, pageable);
+		Page<StudentCouncilPost> posts = (category == null)
+			? postRepository.findAll(pageable)
+			: postRepository.findAllByCategory(category, pageable);
 
-        return posts.map(post ->
-                StudentCouncilPostMapper.toListItem(post, currentUserId)
-        );
-    }
+		return posts.map(post ->
+			StudentCouncilPostMapper.toListItem(post, currentUserId)
+		);
+	}
 
-    @Transactional
-    public void delete(Long councilId, Long postId) {
+	@Transactional
+	public void delete(Long councilId, Long postId) {
 
-        StudentCouncilPost post = postRepository.findByIdWithWriter(postId)
-                .orElseThrow(PostNotFoundException::new);
+		StudentCouncilPost post = postRepository.findByIdWithWriter(postId)
+			.orElseThrow(PostNotFoundException::new);
 
-        if (!post.getWriter().getId().equals(councilId)) {
-            throw new NotPostWriterException();
-        }
+		if (!post.getWriter().getId().equals(councilId)) {
+			throw new NotPostWriterException();
+		}
 
-        List<PostImage> postImages = postImageRepository.findAllByPost(post);
+		List<PostImage> postImages = postImageRepository.findAllByPost(post);
 
-        List<String> deleteTargets = new ArrayList<>();
+		List<String> deleteTargets = new ArrayList<>();
 
-        if (post.getThumbnailImageUrl() != null) {
-            deleteTargets.add(post.getThumbnailImageUrl());
-        }
+		if (post.getThumbnailImageUrl() != null) {
+			deleteTargets.add(post.getThumbnailImageUrl());
+		}
 
-        postImages.stream()
-                .map(PostImage::getImageUrl)
-                .forEach(deleteTargets::add);
+		postImages.stream()
+			.map(PostImage::getImageUrl)
+			.forEach(deleteTargets::add);
 
-        postImageRepository.deleteAll(postImages);
-        postRepository.delete(post);
+		postImageRepository.deleteAll(postImages);
+		postRepository.delete(post);
 
-        for (String imageUrl : deleteTargets) {
-            try {
-                presignedUrlService.deleteImage(imageUrl);
-            } catch (Exception e) {
-                log.warn("OCI 파일 삭제 실패: {}", imageUrl);
-            }
-        }
-    }
+		for (String imageUrl : deleteTargets) {
+			try {
+				presignedUrlService.deleteImage(imageUrl);
+			} catch (Exception e) {
+				log.warn("OCI 파일 삭제 실패: {}", imageUrl);
+			}
+		}
+	}
 
-    @Transactional
-    public PostResponseDto update(
-            Long councilId,
-            Long postId,
-            PostRequestDto dto
-    ) {
+	@Transactional
+	public PostResponseDto update(
+		Long councilId,
+		Long postId,
+		PostRequestDto dto
+	) {
 
-        if (dto.imageUrls() != null && dto.imageUrls().size() > 10) {
-            throw new PostImageLimitExceededException();
-        }
+		if (dto.imageUrls() != null && dto.imageUrls().size() > 10) {
+			throw new PostImageLimitExceededException();
+		}
 
-        StudentCouncilPost post = postRepository.findByIdWithFullInfo(postId)
-                .orElseThrow(PostNotFoundException::new);
+		StudentCouncilPost post = postRepository.findByIdWithFullInfo(postId)
+			.orElseThrow(PostNotFoundException::new);
 
-        if (!post.getWriter().getId().equals(councilId)) {
-            throw new NotPostWriterException();
-        }
+		if (!post.getWriter().getId().equals(councilId)) {
+			throw new NotPostWriterException();
+		}
 
-        // 썸네일 검증
-        if (dto.thumbnailImageUrl() == null && dto.thumbnailIcon() == null) {
-            throw new ThumbnailRequiredException();
-        }
+		// 썸네일 검증
+		if (dto.thumbnailImageUrl() == null && dto.thumbnailIcon() == null) {
+			throw new ThumbnailRequiredException();
+		}
 
-        NormalizedDateTime normalized = dto.category().validateAndNormalize(dto);
+		NormalizedDateTime normalized = dto.category().validateAndNormalize(dto);
 
-        String oldThumbnailUrl = post.getThumbnailImageUrl();
-        List<PostImage> oldImages = postImageRepository.findAllByPost(post);
+		String oldThumbnailUrl = post.getThumbnailImageUrl();
+		List<PostImage> oldImages = postImageRepository.findAllByPost(post);
 
-        post.update(
-                dto.title(),
-                dto.content(),
-                dto.place(),
-                normalized.startDateTime(),
-                normalized.endDateTime(),
-                dto.thumbnailImageUrl(),
-                dto.thumbnailIcon(),
-                dto.category()
-        );
+		post.update(
+			dto.title(),
+			dto.content(),
+			dto.place(),
+			normalized.startDateTime(),
+			normalized.endDateTime(),
+			dto.thumbnailImageUrl(),
+			dto.thumbnailIcon(),
+			dto.category()
+		);
 
-        postImageRepository.deleteByPost(post);
+		postImageRepository.deleteByPost(post);
 
-        if (dto.imageUrls() != null) {
-            for (String imageUrl : dto.imageUrls()) {
-                postImageRepository.save(
-                        StudentCouncilPostMapper.toEntity(post, imageUrl)
-                );
-            }
-        }
+		if (dto.imageUrls() != null) {
+			for (String imageUrl : dto.imageUrls()) {
+				postImageRepository.save(
+					StudentCouncilPostMapper.toEntity(post, imageUrl)
+				);
+			}
+		}
 
-        cleanupUnusedImages(oldThumbnailUrl, oldImages, dto);
+		cleanupUnusedImages(oldThumbnailUrl, oldImages, dto);
 
-        List<String> imageUrls = postImageRepository
-                .findAllByPostOrderByIdAsc(post)
-                .stream()
-                .map(PostImage::getImageUrl)
-                .toList();
+		List<String> imageUrls = postImageRepository
+			.findAllByPostOrderByIdAsc(post)
+			.stream()
+			.map(PostImage::getImageUrl)
+			.toList();
 
-        return StudentCouncilPostMapper.toDetail(
-                post,
-                imageUrls,
-                councilId
-        );
-    }
+		return StudentCouncilPostMapper.toDetail(
+			post,
+			imageUrls,
+			councilId
+		);
+	}
 
-    //이미지 삭제
-    private void cleanupUnusedImages(
-            String oldThumbnailUrl,
-            List<PostImage> oldImages,
-            PostRequestDto dto
-    ) {
-        List<String> newUrls = dto.imageUrls() == null
-                ? List.of()
-                : dto.imageUrls();
+	//이미지 삭제
+	private void cleanupUnusedImages(
+		String oldThumbnailUrl,
+		List<PostImage> oldImages,
+		PostRequestDto dto
+	) {
+		List<String> newUrls = dto.imageUrls() == null
+			? List.of()
+			: dto.imageUrls();
 
-        List<String> deleteTargets = new ArrayList<>();
+		List<String> deleteTargets = new ArrayList<>();
 
-        // 썸네일 변경 시 이전 썸네일
-        if (oldThumbnailUrl != null &&
-                !oldThumbnailUrl.equals(dto.thumbnailImageUrl())) {
-            deleteTargets.add(oldThumbnailUrl);
-        }
+		// 썸네일 변경 시 이전 썸네일
+		if (oldThumbnailUrl != null &&
+			!oldThumbnailUrl.equals(dto.thumbnailImageUrl())) {
+			deleteTargets.add(oldThumbnailUrl);
+		}
 
-        // 본문 이미지 중 제거된 이미지
-        oldImages.stream()
-                .map(PostImage::getImageUrl)
-                .filter(url -> !newUrls.contains(url))
-                .forEach(deleteTargets::add);
-        //삭제
-        for (String imageUrl : deleteTargets) {
-            if (imageUrl == null || imageUrl.isBlank()) {
-                continue;
-            }
+		// 본문 이미지 중 제거된 이미지
+		oldImages.stream()
+			.map(PostImage::getImageUrl)
+			.filter(url -> !newUrls.contains(url))
+			.forEach(deleteTargets::add);
+		//삭제
+		for (String imageUrl : deleteTargets) {
+			if (imageUrl == null || imageUrl.isBlank()) {
+				continue;
+			}
 
-            try {
-                presignedUrlService.deleteImage(imageUrl);
-            } catch (Exception e) {
-                log.warn("OCI 파일 삭제 실패 (파일이 없을 수 있음): {}", imageUrl);
-            }
-        }
-    }
+			try {
+				presignedUrlService.deleteImage(imageUrl);
+			} catch (Exception e) {
+				log.warn("OCI 파일 삭제 실패 (파일이 없을 수 있음): {}", imageUrl);
+			}
+		}
+	}
 }
