+++ src/main/java/com/campus/campus/domain/studentcouncilpost/domain/entity/ThumbnailIcon.java
@@ -1,10 +1,10 @@
 package com.campus.campus.domain.studentcouncilpost.domain.entity;
 
 public enum ThumbnailIcon {
-    CAFE,
-    EVENT,
-    NOTICE,
-    FOOD,
-    SPORTS
+	CAFE,
+	EVENT,
+	NOTICE,
+	FOOD,
+	SPORTS
 }
