+++ src/main/java/com/campus/campus/domain/studentcouncilpost/domain/repository/PostImageRepository.java
@@ -1,16 +1,18 @@
 package com.campus.campus.domain.studentcouncilpost.domain.repository;
 
+import java.util.List;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
 import com.campus.campus.domain.studentcouncilpost.domain.entity.PostImage;
 import com.campus.campus.domain.studentcouncilpost.domain.entity.StudentCouncilPost;
-import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostImageRepository extends JpaRepository<PostImage, Long> {
 
-    List<PostImage> findAllByPost(StudentCouncilPost post);
+	List<PostImage> findAllByPost(StudentCouncilPost post);
 
-    void deleteByPost(StudentCouncilPost post);
+	void deleteByPost(StudentCouncilPost post);
 
-    List<PostImage> findAllByPostOrderByIdAsc(StudentCouncilPost post);
+	List<PostImage> findAllByPostOrderByIdAsc(StudentCouncilPost post);
 }
