+++ src/main/java/com/campus/campus/global/oci/application/dto/response/PresignedUrlResponseDto.java
@@ -3,11 +3,12 @@ package com.campus.campus.global.oci.application.dto.response;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 public record PresignedUrlResponseDto(
-        @Schema(description = "Presigned PUT 업로드 URL")
-        String uploadUrl,
+	@Schema(description = "Presigned PUT 업로드 URL")
+	String uploadUrl,
 
-        @Schema(description = "업로드 완료 후 접근 가능한 이미지 URL")
-        String imageUrl
-) {}
+	@Schema(description = "업로드 완료 후 접근 가능한 이미지 URL")
+	String imageUrl
+) {
+}
