--- src/main/java/com/campus/campus/global/oci/exception/OciObjectCopyFailException.java
@@ -1,9 +1,0 @@
-package com.campus.campus.global.oci.exception;
-
-import com.campus.campus.global.common.exception.ApplicationException;
-
-public class OciObjectCopyFailException extends ApplicationException {
-    public OciObjectCopyFailException() {
-        super(ErrorCode.OCI_OBJECT_COPY_FAILED);
-    }
-}
