+++ src/main/java/com/campus/campus/global/oci/exception/OciObjectDeleteFailException.java
@@ -3,7 +3,7 @@ package com.campus.campus.global.oci.exception;
 import com.campus.campus.global.common.exception.ApplicationException;
 
 public class OciObjectDeleteFailException extends ApplicationException {
-  public OciObjectDeleteFailException() {
-    super(ErrorCode.OCI_OBJECT_DELETE_FAILED);
-  }
+	public OciObjectDeleteFailException() {
+		super(ErrorCode.OCI_OBJECT_DELETE_FAILED);
+	}
 }
