+++ src/main/java/com/campus/campus/global/oci/exception/OciPresignedUrlCreateFailException.java
@@ -3,8 +3,7 @@ package com.campus.campus.global.oci.exception;
 import com.campus.campus.global.common.exception.ApplicationException;
 
 public class OciPresignedUrlCreateFailException extends ApplicationException {
-
-    public OciPresignedUrlCreateFailException() {
-        super(ErrorCode.OCI_PRESIGNED_URL_CREATE_FAILED);
-    }
+	public OciPresignedUrlCreateFailException() {
+		super(ErrorCode.OCI_PRESIGNED_URL_CREATE_FAILED);
+	}
 }
